+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 Version 6.1.0-dev
-- added CodeCoverageExtension, Reader, IAnnotationsReaderEngine, PhpAttributesEngine, TAssertions::getCounter() to public api
+- added CodeCoverageExtension, Reader, IAnnotationsReaderEngine, PhpAttributesEngine, TAssertions::getCounter(), TestsStartedEvent, TestsFinishedEvent to public api
 
 Version 6.0.1
 - Cobertura code coverage report now shows coverage in traits